+++ lib/router/group/group_admin/group_admin_screen.dart
@@ -213,6 +213,14 @@ class _GroupAdminScreenState extends State<GroupAdminScreen> {
                           context, RouterPath.groupMembers, groupId);
                     },
                   ),
+                  _NavigationRow(
+                    title: localization.Community_Guidelines,
+                    onTap: () {
+                      final groupId = context.read<GroupIdentifier>();
+                      RouterUtil.router(
+                          context, RouterPath.communityGuidelines, groupId);
+                    },
+                  ),
                 ],
               ),
             ),