+++ CHANGELOG.md
@@ -18,6 +18,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Made it easier to dismiss fullscreen images. [#286](https://github.com/verse-pbc/issues/issues/286)
 - Added Cancel button when adding accounts. [#254](https://github.com/verse-pbc/issues/issues/254)
 - Start caching groups metadatas thus improving the loading times. [#261](https://github.com/verse-pbc/issues/issues/261)
+- Added Onboarding Age Verification screen before signup. [#283](https://github.com/verse-pbc/issues/issues/283)
+- Added Community Guidelines to groups. [#264](https://github.com/verse-pbc/issues/issues/264)
+- Fix inconsistent navbar background color. [#305](https://github.com/verse-pbc/issues/issues/305)
 
 ### Internal Changes
 - Added functions to send push notification registration events to our relay. [#137](https://github.com/verse-pbc/plur/pull/137)